Added jogcon detection
Added JogCon detection.
Also changed type 0x03 to be DualShock 2.

JogCon is usefull as it will report digital protocol when just powered on. It will only report the correct protocol after sending commands to enable analog and rumble motors.
This way it can be detected even if it's not in analog mode. (red led off)