Support for the legacy rumble mode
This mode is present on the japanese SCPH-1150 and uses a single motor.
Both DualShock and DualShock2 supports this rumble mode.
Entering config mode will disable this mode on the controller and it will require a power cycle to be used again.
`MOTOR_OLD_1` and `MOTOR_OLD_2` constants can be changed but some bits must be set. I've set a note on this.